Create menu header. Change margins on recipe text
Created a menu header and added to all html docs. This will serve
as a persistent header across the whole website that contains main
navigation buttons. The background of this header is dark red to add
contrast.
Changed margins on recipe text to space out text better as
well as to fix issue where bullet points and numbers on lists would
appear outside recipeText div on mobile devices.
Changed margins on existing headers to be more consistent with spacing.